Refactoring. Deprecation.
Removed own uX/sX types, replaced with uintX_t/intX_t.
Deprecated the math classes, use glm instead!
Added ScreenConfig to Application.
Refactored Stream classes.

Note: Build NOT tested.

+++ utils/src/utils/Stream.cc
@@ -8,31 +8,31 @@ namespace coin {
 namespace {
 
 template<typename T>
-inline void get (u8* buffer, u32 pos, T& out) {
-    const u32 shl = (sizeof(T) - (pos + 1)) * 8;
+inline void get (uint8_t* buffer, size_t pos, T& out) {
+    const size_t shl = (sizeof(T) - (pos + 1)) * 8;
     out = out | (((T) buffer[pos]) << shl);
 }
 
-template<u32 size, typename T>
+template<size_t size, typename T>
 inline void read (Stream& stream, T& out) {
-    u8 buffer [size];
+    uint8_t buffer [size];
     stream.Read (buffer, size);
-    for (u32 i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         get (buffer, i, out);
     }
 }
 
 
 template<typename T>
-inline void set (T in, u8* buffer, u32 pos) {
-    const u32 shr = (sizeof(T) - (pos + 1)) * 8;
-    buffer[pos] = (u8) (in >> shr);
+inline void set (T in, uint8_t* buffer, size_t pos) {
+    const size_t shr = (sizeof(T) - (pos + 1)) * 8;
+    buffer[pos] = (uint8_t) (in >> shr);
 }
 
-template<u32 size, typename T>
+template<size_t size, typename T>
 inline void write (Stream& stream, T in) {
-    u8 buffer [size];
-    for (u32 i = 0; i < size; ++i) {
+    uint8_t buffer [size];
+    for (size_t i = 0; i < size; ++i) {
         set (in, buffer, i);
     }
     stream.Write (buffer, size);
@@ -41,56 +41,56 @@ inline void write (Stream& stream, T in) {
 }
 
 
-void Stream::ReadU8 (u8& out) {
+void Stream::ReadU8 (uint8_t& out) {
     out = 0;
     Read (&out, 1);
 }
 
-void Stream::ReadU16 (u16& out) {
+void Stream::ReadU16 (uint16_t& out) {
     out = 0;
     read<2> (*this, out);
 }
 
-void Stream::ReadU32 (u32& out) {
+void Stream::ReadU32 (uint32_t& out) {
     out = 0;
     read<4> (*this, out);
 }
 
-void Stream::ReadU64 (u64& out) {
+void Stream::ReadU64 (uint64_t& out) {
     out = 0;
     read<8> (*this, out);
 }
 
 void Stream::ReadString (std::string& out) {
-    u16 size; 
+    uint16_t size; 
     ReadU16 (size);
     if (size > 0) {
         out.resize (size);
-        Read ((u8*) &out[0], size);
+        Read ((uint8_t*) &out[0], size);
     }
 }
 
-void Stream::WriteU8 (const u8 in) {
+void Stream::WriteU8 (const uint8_t in) {
     Write (&in, 1);
 }
 
-void Stream::WriteU16 (const u16 in) {
+void Stream::WriteU16 (const uint16_t in) {
     write<2> (*this, in);
 }
 
-void Stream::WriteU32 (const u32 in) {
+void Stream::WriteU32 (const uint32_t in) {
     write<4> (*this, in);
 }
 
-void Stream::WriteU64 (const u64 in) {
+void Stream::WriteU64 (const uint64_t in) {
     write<8> (*this, in);
 }
 
 void Stream::WriteString (const std::string& in) {
-    u16 size = (u16) in.size (); 
+    uint16_t size = (uint16_t) in.size (); 
     WriteU16 (size);
     if (size > 0) {
-        Write ((const u8*) in.data (), size);
+        Write ((const uint8_t*) in.data (), size);
     }
 }
